Injected camera controls from interactionManager to separate component
