+++ src/js/shot-generator/components/Three/InteractionManager.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import React, { useRef, useCallback, useLayoutEffect, useState, useMemo, useEffect } from 'react'
-import { useThree, useFrame } from 'react-three-fiber'
+import { useThree } from 'react-three-fiber'
 import * as THREE from 'three'
 import { useDraggingManager } from '../../hooks/use-dragging-manager'
 import '../../../shared/IK/utils/Object3dExtension'
@@ -12,7 +12,6 @@ import {
     selectBone,
     selectAttachable,
     updateObjects,
-    updateObject,
     
     undoGroupStart,
     undoGroupEnd,
@@ -23,9 +22,8 @@ import {
     getSceneObjects,
 } from '../../../shared/reducers/shot-generator'
 import BonesHelper from '../../../xr/src/three/BonesHelper'
-import CameraControls from '../../CameraControls'
 import throttle from 'lodash.throttle'
-import isUserModel from '../../helpers/isUserModel'
+import CameraControlsComponent from './CameraControlsComponet'
 
 const getIntersectionTarget = intersect => {
   // character
@@ -59,14 +57,12 @@ const getIntersectionTarget = intersect => {
 const InteractionManager = connect(
     state => ({
         activeCamera: getActiveCamera(state),
-        selections: getSelections(state)
     }),
     {
         selectObject,
         selectObjectToggle,
         selectBone,
         updateObjects,
-        updateObject,
         selectAttachable,
         deselectAttachable,
         undoGroupStart,
@@ -80,7 +76,6 @@ const InteractionManager = connect(
     selectBone,
 
     updateObjects,
-    updateObject,
     selectAttachable,
     deselectAttachable,
 
@@ -88,7 +83,6 @@ const InteractionManager = connect(
     undoGroupEnd,
     withState,
     renderData,
-    selections
 }) => {
     const { scene, gl, camera } = useThree()
 
@@ -96,11 +90,13 @@ const InteractionManager = connect(
     const intersectables = useRef()
     const [lastDownId, setLastDownId] = useState()
     const [dragTarget, setDragTarget] = useState()
+    const [pointerDownEvent, setOnPointDown] = useState()
+    const [pointerUpEvent, setOnPointUp] = useState()
+    const [isCameraControlsEnabled, enableCameraControls] = useState(true)
     const { prepareDrag, drag, updateStore, endDrag } = useDraggingManager(false)
     const gpuPickerInstance = useRef(null)
     const raycaster = useRef(new THREE.Raycaster())
     const mousePosition = useRef(new THREE.Vector2())
-    const cameraControlsView = useRef()
 
     const takeSceneObjects = useCallback(() => {
       let sceneObjects 
@@ -120,61 +116,10 @@ const InteractionManager = connect(
 
 
     useEffect(() => {
-
-      if(!cameraControlsView.current) return
-      cameraControlsView.current.dispose()
-      cameraControlsView.current.domElement = activeGL.domElement
-      cameraControlsView.current.intializeEvents()
       SGIkHelper.getInstance().changeDomElement(activeGL.domElement)
       if(!gpuPickerInstance.current) return
       gpuPickerInstance.current.renderer = activeGL
     }, [activeGL])
-    
-    const updatedCameraObject = useRef(null)
-
-    const onCameraUpdate = ({active, object}) => {
-      if (camera.userData.locked) {
-        return false
-      }
-
-      if (!active) {
-        updateObject(camera.userData.id, {
-          x: object.x,
-          y: object.y,
-          z: object.z,
-          rotation: object.rotation,
-          tilt: object.tilt,
-          roll: object.roll,
-          fov: object.fov
-        })
-      } else {
-        camera.position.x = object.x
-        camera.position.y = object.z
-        camera.position.z = object.y
-        camera.rotation.x = 0
-        camera.rotation.z = 0
-        camera.rotation.y = object.rotation
-        camera.rotateX(object.tilt)
-        camera.rotateZ(object.roll)
-        camera.fov = object.fov
-        camera.isSynchronized = false
-      }
-    }
-
-    useEffect(() => {
-      if(!activeCamera || cameraControlsView.current ) return
-      let sceneObjects = takeSceneObjects()
-
-      cameraControlsView.current = new CameraControls(
-        CameraControls.objectFromCameraState(sceneObjects[activeCamera]),
-        activeGL.domElement,
-        {
-          undoGroupStart,
-          undoGroupEnd,
-          onChange: onCameraUpdate
-        }
-      )
-    }, [activeCamera, takeSceneObjects])
 
     const getGPUPicker = useCallback(() => {
         if(gpuPickerInstance.current === null) {
@@ -182,12 +127,6 @@ const InteractionManager = connect(
         }
         return gpuPickerInstance.current
     }, [])
-
-    const enableCameraControls = (state) => {
-      if(!cameraControlsView.current) return
-      cameraControlsView.current.reset()
-      cameraControlsView.current.enabled = state
-    }
     
     const filterIntersectables = () => {
         intersectables.current = scene.__interaction
@@ -224,55 +163,20 @@ const InteractionManager = connect(
         return intersects
     }  
 
-    const setCameraControlTarget = (selections) => {
-      if(selections.length === 1 && selections[0] === activeCamera) return
-      let selectedObjects = scene.__interaction.filter(object => object.userData.type !== 'camera' && object.userData.type !== 'volume' 
-                                                        && selections.includes(object.userData.id) )
-      if(!selectedObjects.length) {
-        cameraControlsView.current.Target = null
-        return
-      }
-      let target = new THREE.Vector3()
-      for(let i = 0; i < selectedObjects.length; i++) {
-        let selectedObject = selectedObjects[i]
-        if(selectedObject.userData.type === "character") {
-          if(!isUserModel(selectedObject.userData.model)) {
-            let skinnedMesh = selectedObject.getObjectByProperty("type", "SkinnedMesh")
-            let bone = skinnedMesh.skeleton.getBoneByName("Head")
-            target.add(bone.worldPosition())
-          } else {
-            let position = selectedObjects[i].worldPosition()
-            position.y = camera.position.y
-            target.add(position)
-          }
-        } else {
-          target.add(selectedObjects[i].worldPosition())
-        }
-      }
-      target.divideScalar(selectedObjects.length)
-      cameraControlsView.current.Target = target
-    }
-
     useEffect(() => {
-      setCameraControlTarget(selections)
-    }, [selections])
-
-    useMemo(() => {
         if(dragTarget){
           let selections = takeSelections()
           let { target, x, y } = dragTarget
-          prepareDrag( target, { x, y, useIcons:true, camera, scene, selections })
           enableCameraControls(false)
+          prepareDrag( target, { x, y, useIcons:true, camera, scene, selections })
           undoGroupStart()
         }
     }, [dragTarget])
 
     const onPointerDown = event => {
-      event.preventDefault()
-      filterIntersectables()
-        let sceneObjects = takeSceneObjects()
+        event.preventDefault()
+        filterIntersectables()
         let selections = takeSelections()
-        cameraControlsView.current.object = CameraControls.objectFromCameraState(sceneObjects[activeCamera])
         // get the mouse coords
         const { x, y } = mouse(event)
         const rect = activeGL.domElement.getBoundingClientRect()
@@ -295,7 +199,8 @@ const InteractionManager = connect(
                 // don't select any bone
                 selectBone(null)
             }
-            cameraControlsView.current.onPointerDown(event)
+            setOnPointDown(event)
+           // cameraControlsView.current.onPointerDown(event)
         } else {
     
             let shouldDrag = false
@@ -404,12 +309,12 @@ const InteractionManager = connect(
               selectBone(null)
               setLastDownId(target.userData.id)
             
-             if (shouldDrag) {
-                setDragTarget({ target, x, y })
-             }
-             else {
-              cameraControlsView.current.onPointerDown(event)
-             }
+            if (shouldDrag) {
+               setDragTarget({ target, x, y })
+            }
+            else {
+              setOnPointDown(event)
+            }
         }
     }
 
@@ -451,6 +356,7 @@ const InteractionManager = connect(
           undoGroupEnd()
         }
         enableCameraControls(true)
+        setOnPointUp(event)
         const selections = takeSelections()
         const sceneObjects = takeSceneObjects()
       
@@ -498,15 +404,6 @@ const InteractionManager = connect(
     
         setLastDownId(null)
     }
-
-    useFrame((state, delta) => {
-      if(cameraControlsView.current) {
-       
-        withState((dispatch, state) => {
-          cameraControlsView.current.update(delta, state)
-        })
-      }
-    }, 0)
     
     useLayoutEffect(() => {
       activeGL.domElement.addEventListener('pointerdown', onPointerDown)
@@ -520,7 +417,15 @@ const InteractionManager = connect(
         window.removeEventListener('pointerup', onPointerUp)
       }
     }, [onPointerDown, onPointerUp, onPointerMove, activeGL])
-    return null 
+
+    return <CameraControlsComponent 
+              pointerDownEvent={ pointerDownEvent }
+              pointerUpEvent={ pointerUpEvent }
+              activeGL={ activeGL }
+              isCameraControlsEnabled={ isCameraControlsEnabled }
+              takeSceneObjects={ takeSceneObjects }
+              activeCamera={ activeCamera }
+              /> 
 }))
 
 export default InteractionManager
